Move AGGREGATION_OPERATION def.
