+++ modules/aggregation-layers/src/index.js
@@ -29,7 +29,7 @@ export {default as _GPUGridLayer} from './gpu-grid-layer/gpu-grid-layer';
 export {default as _NewGridLayer} from './new-grid-layer/new-grid-layer';
 
 export {default as _GPUGridAggregator} from './utils/gpu-grid-aggregation/gpu-grid-aggregator';
-export {AGGREGATION_OPERATION} from './utils/gpu-grid-aggregation/gpu-grid-aggregator-constants';
+export {AGGREGATION_OPERATION} from './utils/aggregation-operation-utils';
 
 import {default as BinSorter} from './utils/bin-sorter';
 import {linearScale, getLinearScale, quantizeScale, getQuantizeScale} from './utils/scale-utils';
